+++ libs/clients/documents-v2/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/documentsClientV2.module'
 export * from './lib/documentsClientV2.config'
 export * from './lib/documentsClientV2.service'
+export * from './lib/documentsListClientV2.service'
 export * from '../gen/fetch/models'
 export {
   CustomersGetDocumentPageSortByEnum,
