+++ libs/clients/documents-v2/src/lib/documentsClientV2.module.ts
@@ -1,8 +1,16 @@
 import { Module } from '@nestjs/common'
-import { DocumentsClientV2Provider } from './documentsClientV2.provider'
+import {
+  DocumentsClientListV2Provider,
+  DocumentsClientV2Provider,
+} from './documentsClientV2.provider'
 import { DocumentsClientV2Service } from './documentsClientV2.service'
+import { DocumentsListClientV2Service } from './documentsListClientV2.service'
 @Module({
-  providers: [DocumentsClientV2Provider, DocumentsClientV2Service],
-  exports: [DocumentsClientV2Service],
+  providers: [
+    DocumentsClientV2Provider,
+    DocumentsClientListV2Provider,
+    DocumentsClientV2Service,
+  ],
+  exports: [DocumentsClientV2Service, DocumentsListClientV2Service],
 })
 export class DocumentsClientV2Module {}
